simplify and seperate ftm related components
