+++ app/src/components/CreateEntity.tsx
@@ -1,9 +1,9 @@
-import React, {PureComponent} from 'react';
+import React, {Component} from 'react';
 import {SelectSchema} from "./ftm/SelectSchema";
 import {GraphLayout, GraphUpdateHandler} from "@alephdata/vislib";
 import {Entity, Schema} from "@alephdata/followthemoney";
 import {Divider} from "@blueprintjs/core";
-import EntityEditor from "./ftm/EntityEditor";
+import {EntityEditor} from "./ftm/EntityEditor";
 
 interface ICreateEntityProps {
   layout: GraphLayout
@@ -15,20 +15,34 @@ interface ICreateEntityState {
   entity?: Entity
 }
 
-export class CreateEntity extends PureComponent<ICreateEntityProps, ICreateEntityState> {
+export class CreateEntity extends Component<ICreateEntityProps, ICreateEntityState> {
   constructor(props: ICreateEntityProps) {
     super(props);
     this.onSchemaSelect = this.onSchemaSelect.bind(this);
+    this.appendToLayout = this.appendToLayout.bind(this)
   }
   state:ICreateEntityState = {}
 
+  appendToLayout(entity:Entity){
+    this.props.layout.appendEntity(entity);
+    this.props.layout.layout();
+    this.props.updateLayout(this.props.layout)
+  }
+
   onSchemaSelect(schema:Schema) {
     this.setState(({entity})=>{
+      // generates a new entity based on selected schema
       const nextEntity = this.props.layout.model.createEntity(schema);
+
+      /**
+       * @description transferring values from old entity to new one where applicable, craiteria is where name and and the type of property matchs
+       * */
       if(entity){
+        // stores properties which has a value
         const nextEntityProps = nextEntity.schema.getProperties()
         entity.getProperties()
           .forEach(property => {
+            // identifies if next and prev properties are similar
             const similarProp = nextEntityProps.has(property.name) && nextEntityProps.get(property.name);
             if(similarProp && (similarProp.type === property.type)){
               // this is the case when we can save the value
@@ -36,7 +50,8 @@ export class CreateEntity extends PureComponent<ICreateEntityProps, ICreateEntit
             }
           })
       }
-      this.props.layout.layout()
+
+
       return ({
         entity: nextEntity
       })
@@ -45,17 +60,20 @@ export class CreateEntity extends PureComponent<ICreateEntityProps, ICreateEntit
 
   render() {
     const {layout, subsequentOf} = this.props;
+    const {entity} = this.state;
+    console.log(entity)
     return <div>
       <SelectSchema
+        disabled={!!entity && !!entity.getProperties().length}
         model={layout.model}
         subsequentOf={subsequentOf}
         onSchemaSelect={this.onSchemaSelect}
       />
       <Divider />
-      {this.state.entity &&  <EntityEditor
-        entity={this.state.entity}
+      {entity &&  <EntityEditor
+        onEntityChanged={this.appendToLayout}
+        entity={entity}
         layout={layout}
-        updateLayout={this.props.updateLayout}
       /> }
     </div>
   }
