+++ app/src/components/ftm/type/common.tsx
@@ -1,0 +1,8 @@
+import {Entity, Property, Values} from "@alephdata/followthemoney";
+
+export interface ITypeProps {
+  values: Values
+  property: Property
+  entity: Entity
+  onPropertyChanged: (values: Values, property: Property) => void
+}
